feat: create persistent service
PersistentStoreable
디코딩, 인코딩된 data 값을 이용하여 Store 클래스에서 CRUD 하는 인터페이스를 정의합니다.
매 테스트마다 데이터를 초기화 하기 위하여 removePersistent 메소드를 포함합니다.

PersistentService
Store 클래스에서 CRUD 동작시 발생할 수 있는 에러와 초기화 실패에 대한 메세지를 정의합니다.

KeychainStore, UserDefaultsStore
PersistentStoreable 프로토콜을 채택하여 CRUD, 데이터 초기화 로직을 구현합니다.
각 Store간 동일한 동작을 위해 UserDefaultsStore의 경우 CRUD시 findObject 메소드를 통해 객체 유무를 확인합니다.

Refs: #29
